# Conflicts:
#	ProductPage.cpp

+++ Cakes.cpp
@@ -12,8 +12,8 @@ float Cakes::FinalPrice() const {
     for (const auto& ornament : ornaments) {
         totalPrice += ornament.getPrice() * ornament.getPieceWeight();
     }
-    if (static_cast<double>(days_fresh) < 2) {
-        totalPrice *= 0.9; // 10% discount
+    if (days_fresh < 2) {
+        totalPrice *= 0.9f; // 10% discount
     }
 
     return totalPrice;
